Converted code and tests to new mock-xp
