+++ package.json
@@ -141,6 +141,7 @@
     "bundle:tests:async": "find test -name '*.test.ts'|cut -d'/' -f2-|npx parallel npx rollup -c rollup.config.tests.js -o build/rollup/test/{.}.js test/{}",
     "check": "npx tsc --noEmit",
     "clean": "rimraf build dist",
+    "cov": "jest --no-cache --coverage",
     "esbuild:parseCsv": "npx esbuild --bundle --external:/lib/cache --external:/lib/explorer* --external:/lib/galimatias --external:/lib/graphql* --external:/lib/guillotine* --external:/lib/http-client --external:/lib/license --external:/lib/router --external:/lib/util/* --external:/lib/xp/* --format=cjs --outbase=src/main/resources --outdir=build/esbuild --tsconfig=tsconfig.esbuild.json src/main/resources/lib/explorer/parseCsv.ts",
     "lint": "npx eslint src/main/resources test",
     "prebuild": "yarn clean",
@@ -149,7 +150,7 @@
     "prewebpack": "(cd packages/@enonic/lib-explorer && yarn build)",
     "profile": "time yarn buildEsEsm && time yarn buildSwc && time yarn buildTs",
     "swc:parseCsv": "npx swc --config-file ./.swcrc.server.json build/esbuild/lib/explorer/parseCsv.js -o build/resources/main/lib/explorer/parseCsv.js",
-    "test": "jest --no-cache --coverage",
+    "test": "jest --no-cache",
     "watch": "webpack -w --stdin",
     "webpack": "webpack"
   }
