+++ src/main/resources/lib/explorer/node/create.test.ts
@@ -7,36 +7,34 @@ import type { sanitize } from '@enonic-types/lib-common';
 
 
 import { ROOT_PERMISSIONS_EXPLORER } from '@enonic/explorer-utils';
-import { JavaBridge } from '@enonic/mock-xp';
-import Log from '@enonic/mock-xp/src/Log';
+import {
+	Log,
+	Server,
+} from '@enonic/mock-xp';
 import {
 	describe,
 	expect,
 	jest,
 	test as it
 } from '@jest/globals';
-import { create } from '../../../../src/main/resources/lib/explorer/node/create';
+import { create } from './create';
 
 
-const log = Log.createLogger({
-	// loglevel: 'debug'
-	loglevel: 'silent'
-});
-// @ts-expect-error TS2339: Property 'log' does not exist on type 'typeof globalThis'.
-global.log = log;
-const javaBridge = new JavaBridge({
-	app: {
-		config: {},
-		name: 'com.enonic.app.explorer',
-		version: '0.0.1-SNAPSHOT'
-	},
-	log
-});
-javaBridge.repo.create({
+const server = new Server({
+	loglevel: 'silent',
+}).createRepo({
 	id: 'com.enonic.app.explorer'
 });
-const writeConnection = javaBridge.connect({
-	branch: 'master',
+
+// eslint-disable-next-line @typescript-eslint/no-namespace
+declare module globalThis {
+	let log: Log
+}
+
+globalThis.log = server.log;
+
+const writeConnection = server.connect({
+	branchId: 'master',
 	repoId: 'com.enonic.app.explorer',
 }) as unknown as WriteConnection;
